requirements.txt
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
oauthlib==3.2.2
pycodestyle==2.10.0
python-dotenv==1.0.0
requests==2.28.2
requests-oauthlib==1.3.1
schedule==1.2.0
tomli==2.0.1
tweepy==4.13.0
urllib3==1.26.15
